requirements.txt
certifi==2019.3.9
SQLAlchemy==1.3.3
wincertstore==0.2
certifi==2019.3.9
mysqlclient==1.4.2.post1
PyMySQL==0.9.3
python-dotenv==0.10.1
SQLAlchemy==1.3.3
wincertstore==0.2
